Extract labels from analytics
