+++ Sources/Analytics/AnalyticsService.swift
@@ -8,5 +8,5 @@ import Foundation
 
 protocol AnalyticsService {
     func start(with configuration: Analytics.Configuration)
-    func trackPageView(title: String, levels: [String], labels: Analytics.Labels?)
+    func trackPageView(title: String, levels: [String], labels: Labels?)
 }
